feat: :truck: Renamed articles to docs
Issue: #100
