+++ src/app/docs/[slug]/page.tsx
